+++ src/types/surfaces/brand.ts
@@ -1,5 +1,0 @@
-export function brand<
-  T extends 'Control' | 'Emitting' | 'Introspection' | 'Monitoring' | 'Subscription'
->(type: T): `@@Strongbus${T}Surface` {
-  return `@@Strongbus${type}Surface`;
-}